VSYM-867 : Fixes for intermittent test failures
- Added logic to continue processing in the
  AWSTaskStatusChecker in case AWS returned an
  InstanceIdNotFoundException soon after provisioning. This is
  a known problem in AWS. Just making our check more forgiving
  and allowing for retries.

- Added time bound checking to the retry logic in the
  AWSInstanceAdapter.  Setting the documentExpiration in the
  test and passing that along to the subtask and the status
  checker and honoring that to bind the number of retries made
  to AWS for getting the machine state both during creation
  and deletion.

- Added a check in the tearDown method to delete instances
  only if provisioning succeeded. There was a case where
  there was a provisoning failure on AWS and the test was
  issuing a delete on compute state with Id "UUID" and AWS
  raising an exception com.amazonaws.AmazonServiceException:
  Invalid id: "88f09121-29a7-41cf-a7e2-933b3b9b1c3d"
  (expecting "i-...") (Service: AmazonEC2; Status Code: 400;
  Error Code: InvalidInstanceID.Malformed; Request ID:
  d2315ec8-db8b-4757-859f-520c512e3fc4)

- Adding a check in the scale test for the maximum account
  limit.  In case the requested number of VMs is more that
  the account limit; then the test is left waiting for a
  response from AWS which throws a quota exception.Adding a
  check and spawning few-er resources to adhere to account
  limits.

- Turning OFF detailed monitoring for VMs provisioned
  from the tests as it was costing money specially when
  running at scale.

+++ photon-model-adapters/aws/src/test/java/com/vmware/photon/controller/model/adapters/awsadapter/TestAWSEnumerationTask.java
@@ -122,13 +122,13 @@ public class TestAWSEnumerationTask extends BasicReusableHostTestCase {
     public static final String TEST_CASE_DELETE_VM = "Delete VM ";
     public static final String TEST_CASE_DELETE_VMS = "Delete multiple VMs ";
     public static final String TEST_CASE_MOCK_MODE = "Mock Mode ";
+    public static final int DEFAULT_TEST_PAGE_SIZE = 5;
 
     public String accessKey = "accessKey";
     public String secretKey = "secretKey";
     public static List<String> testComputeDescriptions = new ArrayList<String>(
             Arrays.asList(zoneId + "~" + T2_NANO_INSTANCE_TYPE,
                     zoneId + "~" + instanceType_t2_micro));
-
 
     @Before
     public void setUp() throws Exception {
@@ -155,7 +155,6 @@ public class TestAWSEnumerationTask extends BasicReusableHostTestCase {
                     new AWSEnumerationAdapterService());
             serviceSelfLinks.add(AWSEnumerationAdapterService.SELF_LINK);
 
-
             // create the compute host, resource pool and the VM state to be used in the test.
             createResourcePoolComputeHostAndVMState();
             waitForServiceStart(host, serviceSelfLinks.toArray(new String[] {}));
@@ -171,8 +170,12 @@ public class TestAWSEnumerationTask extends BasicReusableHostTestCase {
             return;
         }
         try {
-            // Delete all vms from the endpoint that w
+            // Delete all vms from the endpoint that were provisioned from the test.
             host.log("Deleting %d instance created from the test ", instancesToCleanUp.size());
+            if (instancesToCleanUp.size() == 0) {
+                cleanupEC2ClientResources(client);
+                return;
+            }
             deleteAllVMsOnThisEndpoint(host, isMock,
                     outComputeHost.documentSelfLink, instancesToCleanUp);
             // Check that all the instances that are required to be deleted are in
@@ -193,8 +196,8 @@ public class TestAWSEnumerationTask extends BasicReusableHostTestCase {
             // Overriding the page size to test the pagination logic with limited instances on AWS.
             // This is a functional test
             // so the latency numbers maybe higher from this test due to low page size.
-            setQueryPageSize(5);
-            setQueryResultLimit(5);
+            setQueryPageSize(DEFAULT_TEST_PAGE_SIZE);
+            setQueryResultLimit(DEFAULT_TEST_PAGE_SIZE);
             baseLineState = getBaseLineInstanceCount(host, client, testComputeDescriptions);
             host.log(baseLineState.toString());
             // Provision a single VM . Check initial state.
